backend/requirements.txt
# FastAPI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Document Processing
python-docx>=1.1.0
PyMuPDF>=1.23.8

# LLM - Compatible versions
groq>=0.9.0
httpx>=0.27.0

# Configuration
pydantic-settings>=2.1.0

# Async utilities
aiohttp>=3.9.1
aiofiles>=23.2.1

# PDF conversion (optional - requires MS Word or LibreOffice)
docx2pdf>=0.1.8

# Native PDF generation (no external dependencies)
fpdf2>=2.7.0
